src/test/java/com/leetcode/hashtable/medium/ContiguousArray_525Test.java
package com.leetcode.hashtable.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class ContiguousArray_525Test {

  private final ContiguousArray_525 solution = new ContiguousArray_525();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(new int[] {0, 1}, 2), arguments(new int[] {0, 1, 0}, 2));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void findMaxLength(int[] nums, int expectedResult) {
    int actualResult = solution.findMaxLength(nums);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
